Move all Theme related logic to infrastructure/theme.rs
